feat(client) add a composition based client for services
Various Services ( in terms of SDK, not Kong's Service concept)
need to access the underlying client which encapsulates most
of the common functionality required across the library.

This commit adds a client (public) type `Kong`.
All services are instances of `service` type which
is composed of the above client and thus the services can
access the underlying client without any type conversion.