fix(/routes) avoid nil pointer in CreateInService()
The following assignment will result in a nil pointer if `Service` is
nil:
```
r.Service.ID = serviceNameOrID
```

This was an oversight in a previous commit and looking into the Git
history, it seems like this made into master without any tests.

This commit adds a test case for this method and also fix the
description, since only service.ID can be specified.