chore: add comments and rename code
This fixes all the lint warnings in the code currently.

Some of the code is not stable currently and will under go refactoring,
but linting has still been satsified.

The goal here is to have good developer documentation since the library
is going to be used by developers.
Although, at times having a doc line for every function might seem
cumbersome it's well worth it as it forces us to read our own interfaces
and have good documentation for critical functions.

We will be setting up linting on Travis as well.