feat(targets) update targets to new Admin API
Kong 1.0 ports over Upstreams and Targets to new DAO, resulting in a few
breaking changes to Admin API requests/responses.

One test is being skipped currently due to a bug[1] in 1.0.0rc3 release.
This will be enabled once another release candidate with the fix is out.

This change is part of a larger changeset adding support for Kong 1.0.

[1]: https://github.com/Kong/kong/issues/4043