feat: implement reset for all core entities
When `deck reset` is executed, delete all core entities in Kong.
Custom entities are not being handled currently.

Most custom entities in Kong are related to Consumers in Kong, which
will be deleted by a cascade so this shouldn't be a problem (may be a
problem for ACL?).

There could be custom entities in the wild which are not related to
Consumers entity and support for those can be added in future.
Note that go-kong the Admin API binding does support custom entities so
it should be possible.