feat: use a dynamic registry during Solve
A dry run and an actual solve is very similar from the syncer's
perespective. The difference is recorded in the Registry of CUD
operations.

This simplifies the `cmd` package and has a single point of entry for
the code.
This _should_ help in testing as well as another CRUD can be
instantiated and injected for verfication of the DAG that is generated.