feat: implement ping command to verify connectivity with Kong
Often in CI world, one needs a preliminary check command to verify if
the dependent components are up and running or not.

This can also be helpful if deck is being used in a containerized
environement on an orchestration platform where the order in which the
components come up can be unpredictable.