fix: handle transactions properly for go-memdb
This took quite a while to resolve.

The defer should always be abort on error and perform a commit on a
happy path only.
The tests luckily caught this otherwise it would have been a nightmare
to figure this problem out if it happened in the wild.

Tests have been added for most functions in this package.