fix: add dependency on route deletion
If a service is marked for deletion, all it's associated routes should
be deleted first and then the service should be deleted.

This dependency was set up previously but was removed in adb829e due to
an oversight.

The commit adds the dependency back.

The change has been tested manually.

It's high time that we consider coming up with a testing harness and
write tests for the critical part of the library.