feat(diff) drop graph library and use a channel for changes
Generating a DAG based on the changes turned out to be more complicated
than I assumed at the beginning. The sheer number of cases are too
large and writing tests for it was another nightmare that I'd like to
avoid for now.

The current solution uses a channel of events which can be processed by
a pool of goroutines for parallel updates. This should help in the case
where Kong and deck have substantial latencies or the C* cluster backing
Kong takes a while to respond to writes.
This approach introduces a concept of wait or barrier between
interelated changes, for example, if a route is to be created, the
corresponding service should be created else the operation will fail. In
the DAG approach, an edge was added ensuring that the dependency was
satisfied. With the event approach, all create events of services will
be finished and then route events will start.

This might result in small durations when parallel operations although
possible theoritically, will not be performed. This trade-off is not
measured but considered okay as of today. If there are large diffs,
there will be number of operations for specific entities, and
parallelism can be exploited for operations on the entity of the same
type, which should be sufficient in most cases.

Another change is create and update operations are now performed before
delete operations.
This is to cover the following case:
- assume service s1 with route r1 is present in kong
- an update needs to perform: delete s1, create s2, associate r1 to s2

This operation will fail if delete is first performed since s1 can't be
deleted as it has r1 associated with it.

But if s2 is created first, then, r1 is moved from s1 to s2. then s1 can
be deleted safely.

Having said that, it is totally possible to do a Graph based solution
for this if we move to a schema based entity design rather than having a
struct of each Kong entity. A schema based design allows for
abstractions and injections.