feat(solver) simplify solver and rely on postProcess in diff
The contract between solver and diff has changed;
diff now expects the registry to always create/update/delete the
entities and return back for both cases: actual run and a dry run.

This simplifies the logic of updating the local state after an operation
has been performed, making the crud operations very small and simple.

dry solver has been udpated to generate a UUID for create operations and
return the entities as is for update and delete operations.

Currently, the diff.Process() function uses the registry but subsequent
updates will move it out to solver.