feat(state) implement a MD5 sum of string fields indexer
Certificate in Kong doesn't have a name.
It needs to be indexed based on the Cert and Key.
Indexing a concat of entire cert and key can be inefficient for the
underlying radix tree data structure.

An MD5 indexer can be used to index based on other fields in future as
well.

Additionally, the indexers are now moved into a subpackage `indexers`
under `state`.