feat(state) implement Certificate and SNIs storage
SNIs are a part of certificate in Kong and hence are not included as a
separate entity here.

A `GET` on a certificate in Kong returns the associated SNIs as well.
It seems like all SNIs associated with a certificate are returned and
hence even with a large number of SNIs (rarely the case)
associated with the same certificate should not be a problem.
This has been tested with Kong 1.0.