fix(solver) use PUT instead of PATCH for out of sync entities
Kong 1.0.0 introduces PUT support for all entities.
This is useful in cases when certain properties are removed.

For example, consider we have the following route:

```yaml
name: r2
paths:
- /r2
methods:
- GET
- PATCH
preserve_host: false
protocols:
- http
- https
regex_priority: 0
strip_path: true
```

and it is being updated to following, where methods is removed:

```yaml
name: r2
paths:
- /r2
preserve_host: false
protocols:
- http
- https
regex_priority: 0
strip_path: true
```

In such a case, a patch request will not update anything since `methods`
will be empty, meaning it won't be sent along with the requests.
This is a limitation of the `go-kong` library, where if a property is empty
it will omit it in the request sent to Kong. If the library explicitly
sets it to `null` or `[]` for an array, and Kong will respect it.

The limitation in `go-kong` will need to be fixed for the case where passing
a null populates the property with defaults, meaning setting an empty
property is not possible.