fix(state) deepcopy objects before returning them from state
Objects pointers are stored in state, if the same pointer is returned
back to the caller, and caller updates any fields, then it will corrupt
the local state.

This change changes the API of the state package, whereby, the caller
always gets a copy of the object and not the pointer to the object that
is actually stored in the state index.