feat(sovler) output JSON diff for updates
gomesssagediff is a great library but doesn't handle plugin config diffs
and pointers.

Instead, deck now outputs the entire object with a diff of the
properties that have changed. This works well with arrays and we don't
have to worry about pointers.

There is one bug that I'm aware of with this:

```
{ "foo" : [ "bar" ] }
{ "foo" : [ "bar", "bar" ] }
```

When the above two JSONs are diffed, they are diffed and + "bar" is
displayed but the other "bar" which was existing goes away. This
shouldn't be a problem in most cases but if it is, we can fix it in
future.