feat(file) expose API and tag dedup during read
Expose all the types so that it can be used by another application
(kubernetes ingress controller for Kong).

Also, on reading, the common tag is applied to all core entities,
meaning, a user can skip adding the tag.

In future, tag deduplication will be added to the write path as well,
where the common selector tags won't be repeated in every entity.