refactor: cleanup logging
Previously, verbose logging at level 1 didn't output anything.
In a very early version of decK, terraform's DAG implementation was
used which had log statements, without a way of disabling those.
Hence, verbose level 1 was used to present those logs and level 2 was
used for HTTP requests.

With this change, verbose level 1 doesn't do anything, same as before
but if any log statements are added in decK, they will be printed by
default.
This level might be used in future for debug statements.

Level 2 outputs HTTP requests as before.

This is not a breaking change from the user's perspective.