feat(file) add support to read yaml files from a directory
The `-s` flag in diff and sync command now can take a directory path
as input in addition to the previous behavior of taking a single file as
input.
When a directory is provided, decK lists all the files inside the
directory recursively, reads all the YAML files and renders the target
state from it.

See #22