fix: do not inject path by default in service for TCP support
A service of type `tcp`, `tls` (may be `grpc` as well) will require that
the path is not set.
If the path was not set, decK automatically injected `/` as the Path.
There is no way to override whatever decK is auto-injecting.

This changes the behavior of automatic injection and might result in
diffs but it is not a breaking change.

Fix #44