fix(dump) fix compatibility for Kong < 1.3
CACertificate core entitiy was not present in the past
and the Admin API request will error with 404 Not Found for older
versions of Kong.
If we do get a 404 from Kong, instead of erroring out,
return back an empty array of
CACertificates, effectively disabling the entity for versions
which don't have it.
A better solution would be to have a version check, and based
on the version, the entities are loaded and synced.

Fix #59