fix(dump) ignore credentials for consumers not in the sub-set
Credentials in Kong are not tagged at the moment.

This leads to an error condition when only a subset where all credentials
If `select_tags` functionality is used, only a subset of consumers
are loaded in memory and this leads to a case where a credential could
belong to a consumer that is not part of the sub-set.
Such ErrNotFound errors are now ignored, if the select_tags feature is
being used.

The duplication of code here is absolutely shit.

Fix #65