feat: return response object on error
If there is an error, the library returns an opaque err, which is hard
to inspect.
With this change, the http response is returned with the err for the
Do() method.

This solves the problem with Do() but doesn't solve the problem all over
the library where an err during any API call returns an error.
This will likely require a custom error type in the package which
includes the HTTP response.

Also, the error condition has been changed, only requests with Status
codes >= 400 will now return an err.