fix(state) allow the same target associated with different upstreams
A Target with the same IP and port can be associated with multiple
upstreams.
The state table previously had Target set to unique, resulting in unique
key violation.

This change indexed uniquely based on a combination of upstream and
target.

This change affect Kong Ingress Controller as well, which is where this
problem was first noted.

See #60
Co-authored-by: Harry Bagdi <harrybagdi@gmail.com>