feat(docker) optionally base of Kong image
When building the container, the argument `BASE_IMAGE` can be set
to an image name of a Kong image (must be Alpine). In that case
when executing, it will start Kong before executing decK, and stop
it afterwards.

This allows the container to be started with the same environment
variables as any other Kong container, and then decK can access
Kong locally (within the container).

example usecase is a Kong cluster in data-plane mode without an
admin api exposed.