feat(cmd) check for workspace and warn user if one does not exist
decK returns an unhelpful 'Not Found' error if a workspace doesn't
exist.
It now looks up the workspace before a diff or a sync command and
notifies user if one doesn't exist.

decK is not going to automatically create a workspace for three reasons:
- workspace provisioning via an API call vs via UI has different
behavior
- decK is not responsible for managing workspaces themselves but content
inside them
- diff command is a dry-run, creating a workspace inside it is a
violation of that rule