refactor(cmd) change the order of state build
This is in preparation of ID based diff approach, where the current
state needs to be loaded before the target state is constructed.
But, the file also contains the select-tags and workspace info, which is
necessary to build the current state.

Hence, the target is split into two steps, first is reading the file,
and then the second step of building the state from content is done
after the current state is loaded from Kong.