refactor: move state generation to state package
A state can be created using a dump from Kong or a file.
Both can use the same function.
Currently, there are two duplicate codebases to create these states.

This is part 1 of the refactor where the state from a KongRawState is
part of the state package itself.

The file package should reuse state.Get() method again.