feat(solver) show progress in sync mode
This commit is less of a feature and more of a refactor.
The dry package has been removed and instead the entities and operations
are printed to console using the `state.ConsoleString` abstraction.

This reduces the maintainence overhead by quite a bit and further
reduces the difference between diff and sync mode.

This commit is a good example of how the refactoring work on state
package done in the 0.6 cycle is paying itself off.

Breaking Change:

Showing progress during sync command could be labeled as a breaking
change so we need to document it appropriately in the changelog.

Fix #100