feat(file) implement `_plugin_configs` property for config deduplication
Often times, multiple plugins share the same configuration and in such a
case having configuration duplicated across plugins is not very helpful.
In case of a change, all instances need to be updated.

`_plugin_configs` store the configuration of the plugins and they can be
referenced in any plugin via the `_config` parameter.

Q: Why is this feature implemented in code when YAML references can already
achieve the same?
A: To allow for cases when reference needs to exists across multiple
files. Imagine a case where ops introduces the redis connection params
and apps team set the plugin details.

Fix #93