feat: always export ID for CACertificate
mtls-auth credentials can be tied to CACertificate.
This leads to a probelm where mtls-auth credential can contain an ID of
a CACertificate but CACertificate is exported to state file without an
ID. Now, on a subsequent `sync` from scratch, the ID of CACertificate
will be auto generated. This breaks the link between mtls-auth
credential and CACertificate, and the sync fails.

With this patch, like Certificate entity, ID of CACertificate is always
exported, no matter what.