fix(sync) resolve concurrenct wait bug
Background
----------

Syncer works with a single producer, multiple consumers with a queue in
between them which has a constant buffer size of 10:
`const eventBuffer = 10`
Each time the producer ATTEMPTS to add an item into queue, inflightOps
is incremented atomically unconditionally inside queueEvent().
A consumer decrements the counter once it processes an event, no matter
the event was processed successfully or failed.

A consumer's thread (goroutine) exits once the event channel is closed.
This channel is closed by the producer when the producer is done.

Problem
-------

There are two problems:

1. If a producer's queue is full and `stopChan` is closed,
`queueEvent()` incorrectly increments inflightOps and returns an error,
which causes the producer to exit and eventually all consumers to exit.
This causes the inFlightOps to be incremented incorrectly but doesn't
explain the deadlock.
2. If SIGKILL is sent by the user and producer is stuck at `wait()`
because kong is taking time to respond to requests, `wait()` blocks
indefintely and the process appears to be stuck. Given enough time, this
should resolve and the process should eventually exit once all the
timeouts are reached. If this is not the case, we are dealing with yet
another unknown bug.

Solution
--------

Fix problem (1) by decrementing the counter when the counter increment
was not successful. The counter has to be incremented before sending to
the channel otherwise the producer can wrongly conclude that all
outstanding events are complete in `wait()`.

Fix problem (2) by returning from `wait()` if stopChan is closed.

There could be more such deadlocks lurking in the code and we have
simply gotten lucky that they never show up.
In the long-run, we need to refactor the diff.go file for readability
and maintainance using packages such as `github.com/hbagdi/gang`
which simplify goroutine management.

See #198

Co-authored-by: Christoph Wurm <christoph@konghq.com>