fix(builder) reject creds without required fields
Reject credentials without required fields if attempting to perform a
lookup using them.

For many types of credentials, deck attempts to perform a lookup using
some required field if the credential lacks an ID. If this required
field is not actually present, attempting to search using it will result
in a nil dereference panic.

Some credentials that lack these fields are valid if an ID is available,
e.g. for key-auth, creating a credential with no key is valid, as Kong
will automatically generate a key. For others, credentials that lack the
search field are not valid, e.g. basic-auth credentials will not
automatically generate usernames if none is provided. As this behavior
is not consistent across credentials, and some credentials have required
fields other than the search field, this change only reports an error if
deck cannot possibly find an existing credential because it has neither
an ID nor a search field.