fix(state) drop index on target.Target
target.Target represents an IP address or a DNS name in Kong. These can
be same across different Upstreams.
Drop the unique index on Target resource for this reason.

This is similar to the fix applies to ServiceVersion resource in
cf44212.

The test has been updated to remove insertion of Upstream which doesn't
add in value.

Fix #233