fix(dump) add workaround for free mode mtls-auths (#321)
Treat a 403 on the mtls-auth endpoint the same as a 404. This handles
the endpoint's behavior in Enterprise free mode gracefully, whereas
previously the dump would fail entirely.

We eventually want a way to properly determine if an instance is in
Enterprise free mode and avoid unavailable API calls entirely rather
than relying on the return code.