wip: disable stdin/stdout mode for Konnect cmds
Konnect commands work with additional files when handling documents.
This does not work nicely with stdin/stdout mode, as it'd send some
output (the resource manifest/konnect.yaml) to stdout and some (document
contents) to the filesystem. The simplest option to handle this is to
disable this feature for Konnect commands.