feat: introduce usage analytics
Analytics were temporarily removed as we sent them to a third-party
server.
With this new implementation, these are being sent to Kong Inc directly.

Furthermore, to alleviate concerns raised
in https://github.com/Kong/deck/pull/242, we are now providing a flag on
the root-level, in addition to DECK_ANALYTICS environment variable to
tweak the setting.