wip: rework version document dump loop
Rework the loop that retrieves documents associated with service
versions to iterate over version objects in the inner loop rather than
iterating over version indices.

The previous version of this loop could encounter a concurrency bug
where the version passed to ListAllForParent() was not the version in
the outer anonymous function, causing deck to fetch the same document
twice and fail current state building validity checks. The exact cause
of this bug is not clear (we do not modify the service package slice and
specified indices in the anonymous function call), but it does not
appear to occur with the revised loop.