refactor: drop reset package to simplify codebase (#409)
- reset command follows the same code path as diff and sync commands
i.e. entities are exported and reset using the syncer itself, this gets
rid off reset package entirely and minimizes the amount of code that
needs to be maintained
- Previously, we relied on kong's "on-cascade-delete" meaning, only a
single DELETE request was sent to delete a service, which in turn
deleted all the routes of the service and all the plugins on those
routes, services,etc. Now, reset will delete each and every entity in
the correct order (same as diff and sync). This results in a decrease in
performance of reset command as it will take longer. I don't expect this
to be a problem since reset command is used less often and with
concurrent nature of decK, this should not be a burden.