refactor: refactor service entity-specifics into types package
This commit refactors the Service entity diff, sync, post-process logic
into `types` package.
This patch is part of a larger set where we are consolidating the two
tightly coupled `diff` and `solver` packages into a single `diff`
package. As part of the refactor, we are also moving out entity-specific
logic into a separate package that will be injected into diff package as
needed. The intention is to make adding new entities into decK easier
and more maintainable and make the core engine of decK testable and test
it.