refactor: abstract out service diff logic
This commit moves the Create/Update/Delete event generation from diff
package to the `types` package.

This is part of the continued effort to make the diffing engine smaller
and testable since it is a critical piece in the entire codebase of
decK.

This commit moves the logic only for the Service entity to keep the
commit reviewable. The followon commit refactors all the remaining
entities.