fix: drop conditional tagging
When decK v1.7.0 or later is used with select-tag feature and
credentials in state file, decK will error out in most sync runs because
it can't locate the credentials in the current state.
This happens because #282 changed the conditions under which credentials
were exported: only tagged credentials are now exported.

On the other hand, `kongVersion` field in the state builder is not set.
This leads to credentials not being tagged appropriately when they are
read from the state files.

This patch unconditionally tags all credentials and ensures that
credentials are created and updated appropriately.