fix: --select-tag handles duplicates in files (#571)
* fix: unexpected error with --select-tag and multiple files

When running sync with multiple files and --select-tag,
deck simply appends the existing _info.select_tags of the
various files. If the same tag(s) are present in the files,
these are appended together and making deck fail because
tags passed via the command flag and the ones coming from
the files are expected to match, but the way files are
merged may cause duplicates, hence the unexpected failure.

This removes duplicates from the _info.select_tags resulted
from the merged files.