feat: add entities defaults from their schemas
Today deck populates entities defaults defining them as part
of the Defaulter utility. This PR introduces the ability
to pull the schema of an entity from the Admin API and to
add its defaults from there.

What this PR doesn't do is change a thing on Defaulter.
A refactoring PR will follow.