feat: implement --dedup-plugin-configs flag for dump
_plugin_configs feature of decK allows users to abstract out configuration
of plugins that are used repeatedly and then reference them in
services/routes/consumers. This helps with de-duplication of configuration
code. While users would like to take advantage of this, most of them cannot
because de-duplicating the configuration in the first place is a cumbersome
and an error-prone process.

This implements such flag and feature for the dump command.