fix: update consumers when custom_id is also set
Right now decK only uses ID and Username for Consumers lookup,
which is problematic because CustomID is also unique and
modifying a Username causes a Consumer re-creation, which
triggers a collision for the CustomID field.

This PR makes sure that CustomID is included in the lookup
loop, so that such Consumers changes can be made in-place.