fix: use escaped strings for services' path
Right now, when a `url` containing an encoded whitespace (`%20`)
is passed to a service, decK unwraps it and replace it with an
actual whitespace. This is because decK uses `url.Parse` under
the cover, which is the one making this translation.

This commit makes decK URL unwrapping method to encode back
decoded whitespaces.