fix: remove all selected tags when using dump
The expected behaviour for `deck dump` when using `--select-tag`
is the following:
1. selected tags information should be summarized at the top of
   the configuration file inside the `_info` structure
2. selected tags must be removed from entity-level tags

Right now decK doesn't perform (2) consistently, stripping out
selected tags from some entities, but not from some others.
For example:

```
$ deck dump --select-tag test-tag --yes

$ cat kong.yaml
_format_version: "3.0"
_info:
  defaults: {}
  select_tags:
  - test-tag
services:
- connect_timeout: 60000
  enabled: true
  host: mockbin.org
  name: svc1
  port: 80
  protocol: http
  read_timeout: 60000
  retries: 5
  routes:
  - https_redirect_status_code: 301
    name: r1
    path_handling: v0
    paths:
    - /r1
    preserve_host: false
    protocols:
    - http
    - https
    regex_priority: 0
    request_buffering: true
    response_buffering: true
    strip_path: true
    tags:
    - another-test-tag
  tags:
  - test-tag
  write_timeout: 60000
```

Here, `test-tag` is correctly stripped out from the route
entity, which now has only the `another-test-tag` tag, but the
`test-tag` is still included in the service entity.

This PR makes sure that selected tags are removed from all
entities, so that this information will only be included in the
`_info` structure.