refactor(diff): allow custom printer funcs
when using deck as a lib, if we want to capture the operations logs (creations, updates, deletions), we need to reconfigure the global vars in the `cprint` package.
this is an issue if we want to do different deck operations in parallel.

this change introduce new parameters to the Syncer, to configure custom printer funcs, without depending on the `cprint` package.
Note that by default if no custom printer funcs are configured, the ones from the `cprint` package will be used - so no behaviour change.