feat: enhance error messages during dump for plugins
This update adds additional information if a failure occurs when
gathering referenced (or parent) resources; consumer, service, and
route. The error message will contain the ID of the referenced resource
that failed to be retrieved and include the plugin ID and name for
easily identifying where the missing resource is referenced.