fix: set default strip_path to true
In the past decK had to set the default value of the `strip_path`
field to [false](https://github.com/Kong/deck/blob/v1.26.0/utils/constants.go#L22)
because of some limitation in the library used for defaults injection:
this was a workaround to allow this field to be set to `false`.
Then, if the field was not set in the configuration, the field was then
turned into `true` at runtime via inspection of the Kong Admin API.

This commit changes the decK default for the `strip_path` field to `true`,
since now the underlying library has fixed the issue that triggered
this workaround in the first place.