fix: handle tags for old consumer-groups implementation
Tags support for Consumer Groups was introduced in 3.2.
Before that, making a request with a tag would have no
effect in the Gateway:

```

$ http :8001/plugins\?tags=foo
HTTP/1.1 200 OK
{
    "data": [],
    "next": null
}

$ http :8001/consumer_groups\?tags=foo
HTTP/1.1 200 OK
{
    "data": [
        {
            "created_at": 1699337744,
            "id": "d241ee7f-875a-41f1-84fb-c509df9e716e",
            "name": "my-cg"
        }
    ],
    "next": null
}
```

For this reason, when a Consumer Group is present and a `sync` with
`select_tags` is ran, the existing Consumer Group is wiped out.

This commit introduces a custom way to handle tags for Consumer Groups
when the `select_tags` is present in the input. This doesn't mean
we are implementing tags support for older versions of the Gateway,
but it means that if a Consumer Group (without a tag) exists, it won't
be wiped out when a `sync` with `select_tags` is ran.