fix: use hardcoded Konnect version instead of retrieving it
This commit removes the logic to fetch the Konnect version
from Konnect API and it simply hardcodes it. The Konnect version
is used only to evaluate and decide which format_version to expect
or dump in state files. Since Konnect only uses '_format_verion: 3.0',
and since Konnect always uses the latest Kong version, then we
can simply hardcode the Konnect version and avoid an unnecessary
network call