DELETE THIS BRANCH: modifying decK to work with port forward Konnect
This was an attempt to force consumer groups to work correctly with
Konnect since the translator endpoints contain a bug. Changes to
go-database-reconciler (locally) were also attempted with isKonnect and
various Konnect API calls which are not available in port forwarding.

port-forward-updates.patch is the attempt I had - tis utter garbage.