fix: validate presence of flags and args in file patch command
The command intends to patch input files either via selector-value flags or patch file arguments.
The change ensures that at least one of these is present, but not both at the same time.
Till now, the command accepted patches from flags and patch files at the same time,
but ignored the patches passed via flags. This was a confusing behaviour for customers.

Fix #1130