fix: default konnect url with konnect addr flag not accounted for.
Previously, if a deck command was issued with --konnect-addr flag
but with the default konnectUrl https://us.api.konghq.com, the
flag value was getting ignored as deck could not figure out
if it was passed by flag or not. This change ensures that if the
flag is set, deck commands are issued against Konnect.

Fixes: https://github.com/Kong/deck/issues/1385