Remove quoted type constraints (#1412)
Type constraints in quotes was required in Terraform 0.11 and earlier, but that form is now deprecated and will be removed in a future version of Terraform. For example:

variable "control_plane_id" {
  type = "string"
  default = "YOUR_CONTROL_PLANE_ID"
}
should be

variable "control_plane_id" {
  type = string
  default = "YOUR_CONTROL_PLANE_ID"
}
if the quotes are present, an error is shown during terraform apply.