feat(apply): add deck gateway apply command
The `deck gateway apply` command allows you to apply partial
configuration to a running Gateway instance.

To do this, it runs a `sync` with the NoDeletes flag enabled.
This means that only new and existing resources are updated.
Existing resources that do not exist in the declarative
configuration file are left untouched.