fix: fixing consumer-group related dump config for deck.
With changes added in go-kong, to skip
listing consumers while doing a GET call
on a consumer_group, dump config with gw
3.9+ starting showing issues. Associations
between a consumer and consumer-group were
missing. Thus, we are pushing the no listing
consumers feature behind a flag for now,
applicable only for dump command as a
starting point.