feat: not adding custom-entity support for konnect.
Gating custom-entities to Kong GW.
Using them with Konnect requires
some work for dumping degraqphl_routes
properly. At the moment, a new route
is created on each sync as Konnect
API for GET /degraphql_routes either
does not exist or is erroing out.
Some changes are needed for creates
too, where Konnect can not process
a degraphql_route if it passes a
service name, along with service id.
Until these limitations are resolved,
we are adding limitations in deck
so as to not allow users to create
custom-entities via Konnect.
Necessary changes are added to sync,
diff and validate commands.