feat: added --skip-consumers-with-consumer-groups to sync and diff commands (#1545)
* feat: added --skip-consumers-with-consumer-groups to sync and diff commands
This flag was earlier enabled on dump command.
We are adding it to sync and diff commands too
as it provides performance benefits in case an
enterprise has thousands of consumers per group.
It ensures that list_consumers=false while
querying a consumer-group. This is beneficial for
enterprise customers who sync customers separately
on a different cadence but require frequent changes
in other entity configurations. Not listing consumers
leads to a boost in performance in such cases.

* tests: added Konnect test; changed testdata file

* tests: fixed Konnect test - adding missing config

* fix: added a confirmation prompt for the case where select-tag is missing;

* style: added clarifying comment for flag usage with Konnect

* fix: fixed confirmation prompt

* feat: added --yes bool flag for sync to run non-interactively in case of prompts

* style: removed duplicate utility function

* fix: added --yes flag to diff command